device-posture/device-posture-change-event.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
'use strict';

promise_test(async (t) => {
  const watcher = new EventWatcher(t, navigator.devicePosture, ['change']);
  const postures = ['folded', 'continuous', 'folded'];
  for (const posture of postures) {
    await Promise.all([
      watcher.wait_for('change'),
      test_driver.set_device_posture(posture)
    ]);
    assert_equals(posture, navigator.devicePosture.type);
  }
}, 'Tests the Device Posture API change event handler.');
</script>
